fix/V1-177 - the types are placed in a separate file
