--- frontend/src/components/Layout/Header/SearchCourses/SearchResult/types.ts
@@ -1,4 +1,0 @@
-export interface IClientCourseIds {
-  courseId: string;
-  clientCourseId: string;
-}
